Abstract logic into a concern
