+++ back/app/controllers/web_api/v1/permissions_controller.rb
@@ -1,6 +1,8 @@
 # frozen_string_literal: true
 
 class WebApi::V1::PermissionsController < ApplicationController
+  include LockedUserCustomFieldsConcern
+  
   before_action :set_permission, only: %i[show update reset requirements schema custom_fields access_denied_explanation]
   skip_before_action :authenticate_user
 
@@ -146,30 +148,7 @@ class WebApi::V1::PermissionsController < ApplicationController
     end
   end
 
-  def verification_service
-    @verification_service ||= Verification::VerificationService.new
-  end
 
-  def jsonapi_serializer_params_with_locked_fields
-    constraints = build_locked_custom_fields_constraints
-    jsonapi_serializer_params({ constraints: constraints })
-  end
-
-  def build_locked_custom_fields_constraints
-    return {} unless current_user
-
-    locked_custom_field_keys = verification_service.locked_custom_fields(current_user).map(&:to_s)
-    constraints = {}
-    
-    # Find the corresponding codes for the locked custom field keys
-    # The serializer expects constraints keyed by 'code', not 'key'
-    CustomField.where(key: locked_custom_field_keys).each do |field|
-      code_key = field.code&.to_sym || field.key&.to_sym
-      constraints[code_key] = { locked: true } if code_key
-    end
-    
-    constraints
-  end
 end
 
 # WebApi::V1::PermissionsController.prepend(Verification::Patches::WebApi::V1::PermissionsController)
